create ring_pattern_utils.py
